+++ src/components/RawViewer.vue
@@ -13,7 +13,7 @@
           <q-tab-panels v-model="tabModel" class="bg-grey-9 scroll" style="height: calc(100% - 48px)">
             <q-tab-panel v-for="(item, key) in config" :name="key" :key="`tab-pane-${key}`">
               <template v-if="item.data">
-                <div style="font-size: 1.1rem" class="text-center q-mb-sm text-white" :class="[item.data._color]" v-if="item.description">{{item.description}}</div>
+                <div style="font-size: 1.1rem" class="text-center q-mb-sm text-white" :class="[item.data['x-flespi-color']]" v-if="item.description">{{item.description}}</div>
                 <component v-if="item.wrapper && typeof item.wrapper === 'object'" :is="item.wrapper" :data="getData(item.data)" :inverted="inverted"/>
                 <component v-else-if="item.wrapper && typeof item.wrapper === 'string'" :is="item.wrapper" :class="{'text-white': inverted !== undefined}">{{getData(item.data)}}</component>
                 <div v-else :class="{'text-white': inverted !== undefined}">{{getData(item.data)}}</div>
@@ -69,7 +69,7 @@ export default {
         return data
       } else {
         return Object.keys(data).reduce((result, key) => {
-          if (key === '_description' || key === '_color') { return result }
+          if (key === 'x-flespi-description' || key === 'x-flespi-color') { return result }
           result[key] = data[key]
           return result
         }, {})