+++ src/pages/viewer/Cdns.vue
@@ -80,7 +80,7 @@
       :item="selectedItem"
       :limit="limit"
       originPattern="storage/cdns/:id"
-      :enitity-name="entityName"
+      :entity-name="entityName"
       :isEnabled="true"
       :config="config.logs"
       :style="{height: `calc(100vh - ${isVisibleToolbar ? '100px' : '50px'})`, position: 'relative', ...panelsWidgetsStyle}"