+++ src/pages/viewer/Plugins.vue
@@ -79,7 +79,7 @@
       ref="logs"
       v-if="isInit && active"
       :item="selectedItem"
-      :enitity-name="entityName"
+      :entity-name="entityName"
       :limit="limit"
       originPattern="gw/plugins/:id"
       :isEnabled="true"