+++ src/pages/viewer/Streams.vue
@@ -83,7 +83,7 @@
       :item="selectedItem"
       :limit="limit"
       originPattern="gw/streams/:id"
-      :enitity-name="entityName"
+      :entity-name="entityName"
       :isEnabled="true"
       :config="config.logs"
       :style="{height: `calc(100vh - ${isVisibleToolbar ? '100px' : '50px'})`, position: 'relative', ...panelsWidgetsStyle}"